requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
pillow>=9.0.0
pyyaml>=6.0
tqdm>=4.62.0
tensorboard>=2.8.0
scikit-image>=0.19.0
opencv-python>=4.5.0
einops>=0.6.0
